+++ frontend/app/src/components/home/ReplyingTo.svelte
@@ -1,10 +1,5 @@
 <script lang="ts">
-    import type {
-        EnhancedReplyContext,
-        CreatedUser,
-        OpenChat,
-        ChatIdentifier,
-    } from "openchat-client";
+    import type { EnhancedReplyContext, CreatedUser, OpenChat } from "openchat-client";
     import { _ } from "svelte-i18n";
     import { rtlStore } from "../../stores/rtl";
     import { createEventDispatcher, getContext } from "svelte";
@@ -19,7 +14,6 @@
     export let replyingTo: EnhancedReplyContext;
     export let user: CreatedUser;
     export let readonly: boolean;
-    export let chatId: ChatIdentifier;
 
     $: me = replyingTo.sender?.userId === user?.userId;
 
@@ -48,8 +42,9 @@
     </h4>
     <ChatMessageContent
         {readonly}
-        {chatId}
+        messageContext={replyingTo.sourceContext}
         fill={false}
+        failed={false}
         {me}
         intersecting={true}
         messageId={replyingTo.messageId}