+++ frontend/app/src/components/home/VisibilityControl.svelte
@@ -6,7 +6,7 @@
     import Radio from "../Radio.svelte";
     import { createEventDispatcher, getContext } from "svelte";
     import Button from "../Button.svelte";
-    import DisappearingMessages from "./DisappearingMessages.svelte";
+    import DurationPicker from "./DurationPicker.svelte";
     import AccessGateControl from "./AccessGateControl.svelte";
 
     type T = $$Generic;
@@ -133,9 +133,9 @@
             </div>
             <div class="info">
                 {#if disappearingMessages}
-                    <DisappearingMessages
-                        {canEditDisappearingMessages}
-                        bind:ttl={candidate.eventsTTL} />
+                    <DurationPicker
+                        disabled={!canEditDisappearingMessages}
+                        bind:milliseconds={candidate.eventsTTL} />
                 {/if}
             </div>
         </Checkbox>