+++ frontend/openchat-agent/src/services/community/mappers.ts
@@ -133,6 +133,9 @@ export function addMembersToChannelResponse(
     if ("CommunityFrozen" in candid) {
         return CommonResponses.communityFrozen();
     }
+    if ("InternalError" in candid) {
+        return CommonResponses.internalError();
+    }
     throw new UnsupportedValueError(
         "Unexpected ApiAddMembersToChannelResponse type received",
         candid,
@@ -180,6 +183,10 @@ function failedGateCheckReason(candid: ApiGateCheckFailedReason): GateCheckFaile
         console.warn("PaymentFailed: ", candid);
         return "payment_failed";
     }
+    if ("InsufficientBalance" in candid) {
+        return "insufficient_balance";
+    }
+
     throw new UnsupportedValueError("Unexpected ApiGateCheckFailedReason type received", candid);
 }