+++ frontend/openchat-agent/src/services/group/candid/idl.d.ts
@@ -67,7 +67,9 @@ import {
     ConvertIntoCommunityResponse,
     FollowThreadResponse,
     UnfollowThreadResponse,
-    OptionalMessagePermissions
+    OptionalMessagePermissions,
+    AcceptP2PSwapResponse,
+    CancelP2PSwapResponse
 } from "./types";
 export {
     _SERVICE as GroupService,
@@ -138,6 +140,8 @@ export {
     FollowThreadResponse as ApiFollowThreadResponse,
     UnfollowThreadResponse as ApiUnfollowThreadResponse,
     OptionalMessagePermissions as ApiOptionalMessagePermissions,
+    AcceptP2PSwapResponse as ApiAcceptP2PSwapResponse,
+    CancelP2PSwapResponse as ApiCancelP2PSwapResponse
 };
 
 export const idlFactory: IDL.InterfaceFactory;